+++ rx/concurrency/mainloopscheduler/__init__.py
@@ -4,3 +4,4 @@ from .geventscheduler import GEventScheduler
 from .twistedscheduler import TwistedScheduler
 from .tkinterscheduler import TkinterScheduler
 from .pygamescheduler import PyGameScheduler
+from .qtscheduler import QtScheduler